invert control
