+++ dusty/commands/setup.py
@@ -62,15 +62,7 @@ def setup_dusty_config(mac_username=None, specs_repo=None, nginx_includes_dir=No
     config_dictionary = {constants.CONFIG_MAC_USERNAME_KEY: mac_username,
                          constants.CONFIG_SPECS_REPO_KEY: specs_repo,
                          constants.CONFIG_NGINX_DIR_KEY: nginx_includes_dir}
-    return Payload(save_dusty_config_from_setup, config_dictionary)
-
-def save_dusty_config_from_setup(config):
-    """
-    Saves the config values retrieved from the dusty setup command; since this is the first
-    time that mac_username is set, we need to also load the ssh_auth socket at this point
-    """
-    save_dusty_config(config)
-    check_and_load_ssh_auth()
+    return Payload(save_dusty_config, config_dictionary)
 
 def save_dusty_config(config):
     for key, value in config.iteritems():
