+++ dusty/preflight.py
@@ -9,7 +9,7 @@ import logging
 import subprocess
 import warnings
 
-from .config import write_default_config, get_config_value
+from .config import write_default_config, get_config_value, check_and_load_ssh_auth
 from . import constants
 from .warnings import daemon_warnings
 
@@ -70,27 +70,6 @@ def _ensure_run_dir_exists():
 def _ensure_config_dir_exists():
     if not os.path.exists(constants.CONFIG_DIR):
         os.makedirs(constants.CONFIG_DIR)
-
-def check_and_load_ssh_auth():
-    """
-    Will check the mac_username config value; if it is present, will load that user's
-    SSH_AUTH_SOCK environment variable to the current environment.  This allows git clones
-    to behave the same for the daemon as they do for the user
-    """
-    mac_username = get_config_value('mac_username')
-    if not mac_username:
-        logging.info("Can't setup ssh authorization; no mac_username specified")
-    else:
-        user_id = subprocess.check_output(['id', '-u', mac_username])
-        _load_ssh_auth(user_id)
-
-def _load_ssh_auth(user_id):
-    ssh_auth_sock = subprocess.check_output(['launchctl', 'asuser', user_id, 'launchctl', 'getenv', 'SSH_AUTH_SOCK']).rstrip()
-    if ssh_auth_sock:
-        logging.info("Setting SSH_AUTH_SOCK to {}".format(ssh_auth_sock))
-        os.environ['SSH_AUTH_SOCK'] = ssh_auth_sock
-    else:
-        raise RuntimeError("SSH_AUTH_SOCK not determined; git operations may fail")
 
 def preflight_check():
     logging.info('Starting preflight check')
